Preload launch extensions so there is no caveat about e.g. launch_ros
