+++ launch_frontend_py/entity.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 
 """Module for launch_frontend_py Entity class."""
+import builtins
+import keyword
 from collections.abc import Iterable
 from typing import (
     List,
@@ -30,7 +32,26 @@ from launch.utilities.type_utils import (
     is_instance_of,
 )
 
-from .util import is_reserved_identifier
+
+def is_reserved_identifier(name: str) -> bool:
+    """
+    Check if a name is a reserved identifier in Python.
+
+    Used to avoid naming issues in the launch DSL that overlap with Python reserved words.
+    """
+    return keyword.iskeyword(name) or name in dir(builtins)
+
+
+def make_valid_name(name: str) -> str:
+    """
+    Make a valid Python identifier for an action or attribute name.
+
+    If the name is a reserved identifier in Python, append an underscore to it.
+    """
+    if is_reserved_identifier(name):
+        return f'{name}_'
+    return name
+
 
 
 class Entity(BaseEntity):
@@ -108,9 +129,7 @@ class Entity(BaseEntity):
         See :py:meth:`launch.frontend.Entity.get_attr`.
         Does not apply type coercion, only checks if the read value is of the correct type.
         """
-        # Reserved identifiers are all suffixed with an underscore
-        if is_reserved_identifier(name):
-            name += '_'
+        name = make_valid_name(name)
 
         if name not in self.__attrs:
             if not optional:
