--- launch_frontend_py/util.py
@@ -1,25 +1,0 @@
-# Copyright 2025 Polymath Robotics, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import builtins
-import keyword
-
-
-def is_reserved_identifier(name: str) -> bool:
-    """
-    Check if a name is a reserved identifier in Python.
-
-    Used to avoid naming issues in the launch DSL that overlap with Python reserved words.
-    """
-    return keyword.iskeyword(name) or name in dir(builtins)
